pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[project]
name = "shiroko"
version = "0.0.0-alpha.1"
authors = [{ name = "HumXC", email = "Hum-XC@outlook.com" }]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.7"
dependencies = ["protobuf", "grpcio", "zeroconf"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
]
[tool.hatch.build.targets.wheel.force-include]
"./src/shiroko/protos" = "shiroko/protos"
[project.urls]
"Homepage" = "https://github.com/HumXC/shiroko-python"
"Bug Tracker" = "https://github.com/HumXC/shiroko-python/issues"
